+++ safe_kit/__init__.py
@@ -2,7 +2,9 @@
 from safe_kit.adapter import EthAdapter, Web3Adapter
 from safe_kit.errors import SafeKitError, SafeTransactionError
 from safe_kit.factory import SafeFactory
+from safe_kit.multisend import MultiSend
 from safe_kit.safe import Safe
+from safe_kit.service import SafeServiceClient
 from safe_kit.types import SafeAccountConfig, SafeTransaction, SafeTransactionData
 
 __all__ = [
@@ -15,4 +17,6 @@ __all__ = [
     "Web3Adapter",
     "SafeKitError",
     "SafeTransactionError",
+    "MultiSend",
+    "SafeServiceClient",
 ]